refactor(router): extract qcomponent partial helpers
